users can now specify amount to deposit

+++ metaserver/client.go
@@ -916,7 +916,7 @@ func (client *Client) DeleteContract(renterID string, contractID string) error {
 	return nil
 }
 
-func (client *Client) CreatePaypalPayment(amount int) (string, error) {
+func (client *Client) CreatePaypalPayment(amount float64) (string, error) {
 	if client.token == "" {
 		return "", errors.New("must authorize before calling this method")
 	}
